排名算法

+++ app/Api/Controllers/DownVoteController.php
@@ -7,6 +7,7 @@ use App\Models\Answer;
 use App\Models\Article;
 use App\Models\Comment;
 use App\Models\DownVote;
+use App\Models\Support;
 use App\Services\RateLimiter;
 use Illuminate\Http\Request;
 
@@ -45,13 +46,19 @@ class DownVoteController extends Controller
             throw new ApiException(ApiException::VISIT_LIMIT);
         }
 
+        $support = Support::where("user_id",'=',$loginUser->id)->where('supportable_type','=',get_class($source))->where('supportable_id','=',$source_id)->first();
+        if ($support) {
+            throw new ApiException(ApiException::USER_DOWNVOTE_ALREADY_SUPPORT);
+        }
 
         /*再次踩相当于是取消踩*/
         $downvote = DownVote::where("user_id",'=',$loginUser->id)->where('source_type','=',get_class($source))->where('source_id','=',$source_id)->first();
         if($downvote){
             $downvote->delete();
             $source->decrement('downvotes');
-            return self::createJsonData(true,['tip'=>'取消踩成功','type'=>'cancel_downvote'],ApiException::SUCCESS,'取消踩成功');
+            return self::createJsonData(true,['tip'=>'取消踩成功','type'=>'cancel_downvote',
+                'support_description'=>$source->getSupportRateDesc(),
+                'support_percent'=>$source->getSupportPercent()],ApiException::SUCCESS,'取消踩成功');
         }
 
         $data = [
@@ -66,7 +73,8 @@ class DownVoteController extends Controller
         if($downvote){
             $source->increment('downvotes');
         }
-        return self::createJsonData(true,['tip'=>'踩成功','type'=>'downvote'],ApiException::SUCCESS,'踩成功');
+        return self::createJsonData(true,['tip'=>'踩成功','type'=>'downvote','support_description'=>$source->getSupportRateDesc(),
+            'support_percent'=>$source->getSupportPercent()],ApiException::SUCCESS,'踩成功');
     }
 
 }
